# Conflicts:
#	src/main/java/com/mercadolibre/group8_bootcamp_finalproject/exceptions/OperatorNotInWarehouseException.java
#	src/main/java/com/mercadolibre/group8_bootcamp_finalproject/service/impl/InboundOrderServiceImpl.java
#	src/main/java/com/mercadolibre/group8_bootcamp_finalproject/service/impl/OperatorService.java

+++ src/main/java/com/mercadolibre/group8_bootcamp_finalproject/service/IWarehouseService.java
@@ -5,7 +5,7 @@ import com.mercadolibre.group8_bootcamp_finalproject.model.WarehouseSection;
 
 public interface IWarehouseService {
 
-    WarehouseProductListDTO findAllProductsFromWarehouseById (Integer id);
+    WarehouseProductListDTO findAllProductsFromWarehouseById (Long id);
 
     void verifySectionCapability(WarehouseSection warehouseSection, int quantity);
